+++ android/app/src/main/java/com/lux2983/smartaccountdemo/MainActivity.java
@@ -1,0 +1,5 @@
+package com.lux2983.smartaccountdemo;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}